requirements.txt
langchain.core
openai
requests 
streamlit
langchain.community
langchain.openai 
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
beautifulSoup4
pypdf
chromadb
faiss-cpu
ollama